+++ lib/ng-nest/ui/scrollable/scrollable.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
-import { XScrollableComponent } from './scrollable';
+import { XScrollableComponent } from './scrollable.component';
 
 @NgModule({
   exports: [XScrollableComponent],
   imports: [XScrollableComponent]
 })
-export class XTableViewModule {}
+export class XScrollableModule {}