+++ src/app/app.menus.ts
@@ -205,7 +205,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-72",
+    "id": "1-73",
     "pid": "1",
     "name": "statistic",
     "routerLink": "docs/en_US/components/statistic",
@@ -214,7 +214,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-76",
+    "id": "1-77",
     "pid": "1",
     "name": "table",
     "routerLink": "docs/en_US/components/table",
@@ -223,7 +223,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-78",
+    "id": "1-80",
     "pid": "1",
     "name": "tag",
     "routerLink": "docs/en_US/components/tag",
@@ -232,7 +232,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-79",
+    "id": "1-81",
     "pid": "1",
     "name": "text-retract",
     "routerLink": "docs/en_US/components/text-retract",
@@ -241,7 +241,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-82",
+    "id": "1-84",
     "pid": "1",
     "name": "time-ago",
     "routerLink": "docs/en_US/components/time-ago",
@@ -250,7 +250,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-84",
+    "id": "1-86",
     "pid": "1",
     "name": "time-range",
     "routerLink": "docs/en_US/components/time-range",
@@ -259,7 +259,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-85",
+    "id": "1-87",
     "pid": "1",
     "name": "timeline",
     "routerLink": "docs/en_US/components/timeline",
@@ -268,7 +268,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-88",
+    "id": "1-90",
     "pid": "1",
     "name": "tree",
     "routerLink": "docs/en_US/components/tree",
@@ -277,7 +277,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-89",
+    "id": "1-91",
     "pid": "1",
     "name": "tree-file",
     "routerLink": "docs/en_US/components/tree-file",
@@ -376,7 +376,7 @@ export const menus: AppMenu[] = [
     "category": "Feedback"
   },
   {
-    "id": "1-69",
+    "id": "1-70",
     "pid": "1",
     "name": "skeleton",
     "routerLink": "docs/en_US/components/skeleton",
@@ -385,7 +385,7 @@ export const menus: AppMenu[] = [
     "category": "Feedback"
   },
   {
-    "id": "1-86",
+    "id": "1-88",
     "pid": "1",
     "name": "tooltip",
     "routerLink": "docs/en_US/components/tooltip",
@@ -502,7 +502,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-68",
+    "id": "1-69",
     "pid": "1",
     "name": "select",
     "routerLink": "docs/en_US/components/select",
@@ -511,7 +511,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-71",
+    "id": "1-72",
     "pid": "1",
     "name": "slider-select",
     "routerLink": "docs/en_US/components/slider-select",
@@ -520,7 +520,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-75",
+    "id": "1-76",
     "pid": "1",
     "name": "switch",
     "routerLink": "docs/en_US/components/switch",
@@ -529,7 +529,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-80",
+    "id": "1-82",
     "pid": "1",
     "name": "textarea",
     "routerLink": "docs/en_US/components/textarea",
@@ -538,7 +538,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-83",
+    "id": "1-85",
     "pid": "1",
     "name": "time-picker",
     "routerLink": "docs/en_US/components/time-picker",
@@ -547,7 +547,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-87",
+    "id": "1-89",
     "pid": "1",
     "name": "transfer",
     "routerLink": "docs/en_US/components/transfer",
@@ -556,7 +556,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-90",
+    "id": "1-92",
     "pid": "1",
     "name": "tree-select",
     "routerLink": "docs/en_US/components/tree-select",
@@ -565,7 +565,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-96",
+    "id": "1-98",
     "pid": "1",
     "name": "upload",
     "routerLink": "docs/en_US/components/upload",
@@ -637,7 +637,7 @@ export const menus: AppMenu[] = [
     "category": "Navigation"
   },
   {
-    "id": "1-70",
+    "id": "1-71",
     "pid": "1",
     "name": "slider",
     "routerLink": "docs/en_US/components/slider",
@@ -646,7 +646,7 @@ export const menus: AppMenu[] = [
     "category": "Navigation"
   },
   {
-    "id": "1-73",
+    "id": "1-74",
     "pid": "1",
     "name": "steps",
     "routerLink": "docs/en_US/components/steps",
@@ -655,7 +655,7 @@ export const menus: AppMenu[] = [
     "category": "Navigation"
   },
   {
-    "id": "1-77",
+    "id": "1-79",
     "pid": "1",
     "name": "tabs",
     "routerLink": "docs/en_US/components/tabs",
@@ -988,7 +988,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-72",
+    "id": "1-73",
     "pid": "1",
     "name": "statistic",
     "routerLink": "docs/zh_CN/components/statistic",
@@ -997,7 +997,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-76",
+    "id": "1-77",
     "pid": "1",
     "name": "table",
     "routerLink": "docs/zh_CN/components/table",
@@ -1006,7 +1006,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-78",
+    "id": "1-80",
     "pid": "1",
     "name": "tag",
     "routerLink": "docs/zh_CN/components/tag",
@@ -1015,7 +1015,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-79",
+    "id": "1-81",
     "pid": "1",
     "name": "text-retract",
     "routerLink": "docs/zh_CN/components/text-retract",
@@ -1024,7 +1024,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-82",
+    "id": "1-84",
     "pid": "1",
     "name": "time-ago",
     "routerLink": "docs/zh_CN/components/time-ago",
@@ -1033,7 +1033,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-84",
+    "id": "1-86",
     "pid": "1",
     "name": "time-range",
     "routerLink": "docs/zh_CN/components/time-range",
@@ -1042,7 +1042,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-85",
+    "id": "1-87",
     "pid": "1",
     "name": "timeline",
     "routerLink": "docs/zh_CN/components/timeline",
@@ -1051,7 +1051,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-88",
+    "id": "1-90",
     "pid": "1",
     "name": "tree",
     "routerLink": "docs/zh_CN/components/tree",
@@ -1060,7 +1060,7 @@ export const menus: AppMenu[] = [
     "category": "Data"
   },
   {
-    "id": "1-89",
+    "id": "1-91",
     "pid": "1",
     "name": "tree-file",
     "routerLink": "docs/zh_CN/components/tree-file",
@@ -1159,7 +1159,7 @@ export const menus: AppMenu[] = [
     "category": "Feedback"
   },
   {
-    "id": "1-69",
+    "id": "1-70",
     "pid": "1",
     "name": "skeleton",
     "routerLink": "docs/zh_CN/components/skeleton",
@@ -1168,7 +1168,7 @@ export const menus: AppMenu[] = [
     "category": "Feedback"
   },
   {
-    "id": "1-86",
+    "id": "1-88",
     "pid": "1",
     "name": "tooltip",
     "routerLink": "docs/zh_CN/components/tooltip",
@@ -1285,7 +1285,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-68",
+    "id": "1-69",
     "pid": "1",
     "name": "select",
     "routerLink": "docs/zh_CN/components/select",
@@ -1294,7 +1294,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-71",
+    "id": "1-72",
     "pid": "1",
     "name": "slider-select",
     "routerLink": "docs/zh_CN/components/slider-select",
@@ -1303,7 +1303,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-75",
+    "id": "1-76",
     "pid": "1",
     "name": "switch",
     "routerLink": "docs/zh_CN/components/switch",
@@ -1312,7 +1312,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-80",
+    "id": "1-82",
     "pid": "1",
     "name": "textarea",
     "routerLink": "docs/zh_CN/components/textarea",
@@ -1321,7 +1321,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-83",
+    "id": "1-85",
     "pid": "1",
     "name": "time-picker",
     "routerLink": "docs/zh_CN/components/time-picker",
@@ -1330,7 +1330,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-87",
+    "id": "1-89",
     "pid": "1",
     "name": "transfer",
     "routerLink": "docs/zh_CN/components/transfer",
@@ -1339,7 +1339,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-90",
+    "id": "1-92",
     "pid": "1",
     "name": "tree-select",
     "routerLink": "docs/zh_CN/components/tree-select",
@@ -1348,7 +1348,7 @@ export const menus: AppMenu[] = [
     "category": "Form"
   },
   {
-    "id": "1-96",
+    "id": "1-98",
     "pid": "1",
     "name": "upload",
     "routerLink": "docs/zh_CN/components/upload",
@@ -1420,7 +1420,7 @@ export const menus: AppMenu[] = [
     "category": "Navigation"
   },
   {
-    "id": "1-70",
+    "id": "1-71",
     "pid": "1",
     "name": "slider",
     "routerLink": "docs/zh_CN/components/slider",
@@ -1429,7 +1429,7 @@ export const menus: AppMenu[] = [
     "category": "Navigation"
   },
   {
-    "id": "1-73",
+    "id": "1-74",
     "pid": "1",
     "name": "steps",
     "routerLink": "docs/zh_CN/components/steps",
@@ -1438,7 +1438,7 @@ export const menus: AppMenu[] = [
     "category": "Navigation"
   },
   {
-    "id": "1-77",
+    "id": "1-79",
     "pid": "1",
     "name": "tabs",
     "routerLink": "docs/zh_CN/components/tabs",