+++ src/environments/environment.development.ts
@@ -1,6 +1,6 @@
 export const environment = {
   layout: 'test',
-  version: '20.0.6',
+  version: '20.0.7',
   defaultPage: 'color-picker',
   static: 'https://ngnest.com/static'
 };