+++ src/environments/environment.ts
@@ -1,6 +1,6 @@
 export const environment = {
   layout: 'index',
-  version: '20.0.6',
+  version: '20.0.7',
   defaultPage: 'docs',
   static: 'https://ngnest.com/static'
 };